requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==5.1.2
django-allauth==65.1.0
django-cors-headers==4.6.0
django-filter==24.3
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
pillow==11.0.0
PyJWT==2.10.1
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.1
urllib3==2.2.3
